+++ src/manager/TaskManager.java
@@ -4,6 +4,7 @@ import typetask.Epic;
 import typetask.Subtask;
 import typetask.Task;
 import java.util.List;
+import java.util.TreeSet;
 
 public interface TaskManager {
     Task getTaskByIndex(int id);
@@ -47,4 +48,8 @@ public interface TaskManager {
     List<Subtask> getEpicSubtasks(int id);
 
     void updateEpicStatus(Epic epic);
+
+    TreeSet<Task> getPrioritizedTasksSet();
+
+    void updatePrioritizedTasksSet();
 }
