+++ src/typetask/Subtask.java
@@ -35,7 +35,7 @@ public class Subtask extends Task {
                 "id = " + getId() + ", name = " + getName() + ", description = " +
                 getDescription() + ", status = " + getStatus() + ", duration=" +
                 duration +
-                ", startTime=" + startTime +
+                ", startTime=" + startTime.format(DATE_TIME_FORMATTER) +
                 '}';
     }
 }
